more cleanup before release

+++ rvas-core_v0.1.5/src/main/java/core/commands/Info.java
@@ -22,7 +22,6 @@ package core.commands;
  * 
  * */
 
-import core.backend.Config;
 import core.backend.ServerMeta;
 import core.backend.Utilities;
 
@@ -46,7 +45,7 @@ public class Info implements CommandExecutor {
 		Player player = (Player) sender;
 		if (!player.isOp()) return false;
 			
-		String humanUptime = Utilities.calculateTime((double)ServerMeta.getUptime());
+		String humanUptime = Utilities.calculateTime(ServerMeta.getUptime());
 		
 		TextComponent head = new TextComponent("--- Session Stats ---");
 		head.setColor(ChatColor.GOLD); head.setBold(true);
